requirements.txt
exifread
Pillow
piexif
folium
pytz
geopy
reportlab
pandas
matplotlib
jinja2
timezonefinder